normalized parsing, cleaned up some code
