--- agent/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "agent",
-  "version": "1.0.0",
-  "description": "",
-  "main": "embedding_service.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "commonjs",
-  "dependencies": {
-    "@xenova/transformers": "^2.17.2",
-    "dotenv": "^17.3.1",
-    "groq-sdk": "^1.1.1",
-    "unpdf": "^1.4.0"
-  }
-}
