+++ server/src/config.js
@@ -10,13 +10,15 @@ const connectToMongo = async () => await mongoose.connect(process.env.DARIA_DB_U
 // starts the job scheduler
 const scheduleJobs = () => {
 
+    // Scheduler schedules the Jobs in the jobs array
     const jobs = [new ResumeJob()];
     const scheduler = new Scheduler(jobs);
     scheduler.scheduleJobs();
 };
 
 export const setupApp = async () => {
-    // scheduleJobs();
     await connectToMongo();
     console.log('connected to mongodb');
+    // scheduleJobs();
+    // console.log('jobs scheduled');
 };
