+++ server/src/services/orchestration/rag.orchestration.js
@@ -1,12 +1,86 @@
 import { MixedBreadEmbeddingModel } from "../integration/embedding.integration.js";
-import { GroqLanguageModel } from '../integration/llm.integration.js'
+import { GroqLLM } from '../integration/llm.integration.js'
 import { VectorStore } from "../persistence/vectorStore.persistence.js";
 
 // Make objects
 const Embedder = new MixedBreadEmbeddingModel();
-const LLM = new GroqLanguageModel();
+const LLM = new GroqLLM();
 const Store = new VectorStore();
 
+// prompt for query analysis
+const SYSTEM_PROMPT = `
+You are a senior technical recruiter and career strategist with 15+ years of experience
+hiring for top companies. You have reviewed thousands of resumes and know exactly what
+gets candidates interviews.
+
+## Tone Rules (strictly enforced)
+- Always constructive — frame every weakness as an opportunity to improve
+- No discouraging language, no words like "bad", "poor", "terrible", "wrong", "failed"
+- Be direct but supportive — like a mentor, not a critic
+- Every negative observation MUST be paired with a specific fix
+
+## Output Rules (strictly enforced)
+- Respond ONLY with a valid JSON object — no preamble, no explanation, no markdown
+- Follow the exact schema below — no extra fields, no missing fields
+- All arrays must have at least one item
+- Scores must be integers between 0 and 100
+
+## Response Schema
+{
+  "recruiter_scan": {
+    "first_impression": "string — what stands out in the first 6 seconds",
+    "keep_reading": true or false,
+    "keep_reading_reason": "string — specific reason why or why not"
+  },
+  "fit_score": {
+    "overall": 0-100,
+    "vs_target_role": 0-100,
+    "vs_similar_candidates": 0-100,
+    "breakdown": {
+      "skills_match": 0-100,
+      "experience_match": 0-100,
+      "presentation": 0-100
+    }
+  },
+  "gaps": {
+    "hard_skills": ["string — skill gap + how to address it"],
+    "soft_skills": ["string — soft skill gap + how to address it"],
+    "experience": ["string — experience gap + how to address it"],
+    "ats_flags": ["string — ATS issue + how to fix it"]
+  },
+  "latent_skills": [
+    {
+      "skill": "string",
+      "evidence": "string",
+      "how_to_surface": "string"
+    }
+  ],
+  "recruiter_intelligence": {
+    "what_they_really_want": ["string"],
+    "culture_signals": ["string"],
+    "rejection_triggers": ["string — trigger + how to eliminate it"]
+  },
+  "next_actions": {
+    "immediate_resume_fixes": [
+      {
+        "action": "string",
+        "impact": "high or medium or low",
+        "effort": "high or medium or low"
+      }
+    ],
+    "skills_to_acquire": [
+      {
+        "skill": "string",
+        "timeframe": "string",
+        "resource": "string"
+      }
+    ],
+    "ready_for_now": ["string"],
+    "stepping_stone_roles": ["string"],
+    "target_role_path": ["string"]
+  }
+}
+`;
 // connect services
 await Embedder.initialize();
 
@@ -21,9 +95,7 @@ export const doQuery = async (userQuery, resumeChunks) => {
     // filters = {}
     
     // run the query on the vector store
-    console.log(resumeChunks);
-    const queries = Object.keys(resumeChunks).map( async section => {
-        console.log(section);
+    const queries = Object.keys(resumeChunks).map(async section => {
         return await Store.vectorSearch(
             vectorizedQuery[0],
             5,
@@ -31,14 +103,12 @@ export const doQuery = async (userQuery, resumeChunks) => {
         );
     });
 
-    const results = await Promise.all(queries);
-    console.log(results);
-
-    // Call the LLM
-    LLM.preparePrompt(userQuery, modelContext);
-
+    // wait for all queries to resolve
+    const modelContext = await Promise.all(queries);
+    console.log(modelContext);
+    
     // Run the prompt
-    const response = await LLM.executePrompt();
+    const response = await LLM.executePrompt(SYSTEM_PROMPT, userQuery, modelContext);
 
     return response;
 };
